Fast api/housing_management/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
databases
pydantic
passlib
python-dotenv
alembic